fix: 修复 cache 重大错误 - 判断 aof 文件是否超过配置大小时，应为数值类型导致溢出，最终导致分片逻辑错误

+++ cache/aof/file.go
@@ -139,7 +139,7 @@ func (fop *FileOp) needSplit() bool {
 	}
 
 	// 与最大大小比较（将 MB 转换为字节）
-	return fileInfo.Size() > int64(fop.maxSize*1024*1024)
+	return uint64(fileInfo.Size()) > uint64(fop.maxSize)*1024*1024
 }
 
 // Write 在文件中追加数据，并带有换行符。
